+++ frontend/src/app/page/orders/orders.component.ts
@@ -4,6 +4,7 @@ import { Observable } from 'rxjs';
 import { Order } from 'src/app/model/order';
 import { ITableColumn, ConfigService } from 'src/app/service/config.service';
 import { OrderService } from 'src/app/service/order.service';
+import { ToasterService } from 'src/app/service/toaster.service';
 
 @Component({
   selector: 'app-orders',
@@ -22,6 +23,7 @@ export class OrdersComponent implements OnInit {
     private config: ConfigService,
     private orderService: OrderService,
     private router: Router,
+    private toastr: ToasterService
   ) { }
 
   ngOnInit(): void {
@@ -32,8 +34,26 @@ export class OrdersComponent implements OnInit {
   }
 
   onDeleteOne(order: Order): void {
-    this.orderService.remove(order._id).subscribe(
-      () => this.list$ = this.orderService.getAll()
-      )
+    if (window.confirm('Biztosan törli ezt a megrendelést?')) {
+      this.orderService.remove(order._id).subscribe(
+        () => this.list$ = this.orderService.getAll()
+      );
+      this.toastr.showSuccessWithTimeout(`
+        <table class="table">
+          <thead>
+            <tr>
+              <th>Megrendelés azonosítója</th>
+            </tr>
+          </thead>
+          <tbody>
+            <tr>
+              <td>${order._id}</td>
+            </tr>
+          </tbody>
+        </table>
+        </span>`,
+        "A megrendelés sikeresen törlődött:",
+        5000)
+    }
   }
 }
